requirements.txt
Pillow==9.0.1

streamlit==1.10.0